// src/convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  appUsers: typeof appUsers;
  appointments: typeof appointments;
  auth: typeof auth;
  devSetup: typeof devSetup;
  http: typeof http;
  invoices: typeof invoices;
  leads: typeof leads;
  messages: typeof messages;
  migrations: typeof migrations;
  organizations: typeof organizations;
}>;
// ...
export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;